+++ Owncloud iOs Client/Tabs/SettingTab/Account/AccountsViewController.m
@@ -133,7 +133,7 @@
 
 - (void)tableView:(UITableView *)tableView commitEditingStyle:(UITableViewCellEditingStyle)editingStyle forRowAtIndexPath:(NSIndexPath *)indexPath
 {
-    DLog(@"DELETE!!! %ld", indexPath.row);
+    DLog(@"DELETE!!! %ld", (long)indexPath.row);
     if (editingStyle == UITableViewCellEditingStyleDelete) {
         
         AppDelegate *app = (AppDelegate *)[[UIApplication sharedApplication]delegate];
